viewUptade

+++ src/main/java/fr/eni/tp/chat/bo/Chat.java
@@ -21,7 +21,7 @@ public class Chat {
         return poids;
     }
 
-    public void setPoids(int poids) {
+    public void setPoids(double poids) {
         this.poids = poids;
     }
